Add import to Firefly III functionality
