+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/gocarina/gocsv v0.0.0-20210516172204-ca9e8a8ddea8
 	github.com/pkg/errors v0.9.1
 	github.com/satraul/bca-go v1.0.2
+	github.com/satraul/gofirefly v1.0.0
 	github.com/shibukawa/configdir v0.0.0-20170330084843-e180dbdc8da0
 	github.com/shopspring/decimal v1.2.0
 	github.com/urfave/cli/v2 v2.2.0
