+++ main.go
@@ -1,35 +1,24 @@
 package main
 
 import (
-	"bufio"
 	"context"
 	"encoding/csv"
-	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"log" // TODO Implement https://godoc.org/github.com/apex/log/handlers/cli
 	"net/http"
 	"os"
-	"reflect"
-	"syscall"
 	"time"
 
-	"go.bmvs.io/ynab/api"
-	"go.bmvs.io/ynab/api/account"
-	"go.bmvs.io/ynab/api/category"
 	"go.bmvs.io/ynab/api/transaction"
 
-	"github.com/cnf/structhash"
 	"github.com/gocarina/gocsv"
 	"github.com/satraul/bca-go"
-	"github.com/shopspring/decimal"
 	"go.bmvs.io/ynab"
 
 	"github.com/pkg/errors"
 	"github.com/shibukawa/configdir"
 	"github.com/urfave/cli/v2"
-	"golang.org/x/crypto/ssh/terminal"
 )
 
 var (
@@ -38,15 +27,10 @@ var (
 	configDirs             = configdir.New("satraul", "bca-sync-ynab")
 )
 
-type config struct {
-	BCAUser     string `json:"bcaUser"`
-	BCAPassword string `json:"bcaPassword"`
-	YNABToken   string `json:"ynabToken"`
-}
-
 var (
-	noadjust, delete, noninteractive, nostore, reset, csvFlag bool
-	accountName, budget, password, token, username            string
+	noadjust, delete, noninteractive, nostore, reset, csvFlag                bool
+	accountName, budget, password, token, username, fireflyUrl, fireflyToken string
+	days                                                                     int
 )
 
 func main() {
@@ -135,6 +119,25 @@ func main() {
 				Usage:       "instead of creating ynab transactions, generate a csv",
 				Destination: &csvFlag,
 			},
+			&cli.StringFlag{
+				Name:        "firefly-url",
+				Aliases:     []string{"f"},
+				Usage:       "instead of creating ynab transactions, post to firefly iii url",
+				Destination: &fireflyUrl,
+			},
+			&cli.StringFlag{
+				Name:        "firefly-token",
+				Aliases:     []string{"ft"},
+				Usage:       "firefly token for use with -f / --firefly-url",
+				Destination: &fireflyToken,
+			},
+			&cli.IntFlag{
+				Name:        "days",
+				Aliases:     []string{"n"},
+				Value:       27,
+				Usage:       "fetch transactions from n number of days ago (max is 27)",
+				Destination: &days,
+			},
 		},
 		Action: actionFunc,
 	}
@@ -180,6 +183,10 @@ func actionFunc(c *cli.Context) error {
 		return nil
 	}
 
+	if fireflyUrl != "" {
+		createFireflyTransactions(trxs, ctx)
+	}
+
 	var (
 		yc = ynab.NewClient(config.YNABToken)
 	)
@@ -202,37 +209,16 @@ func actionFunc(c *cli.Context) error {
 	return nil
 }
 
-func getOrDeleteConfig(username string, password string, token string, delete bool, noninteractive bool, reset bool, nostore bool) (*config, error) {
-	var (
-		config = config{BCAUser: username, BCAPassword: password, YNABToken: token}
-		folder = configDirs.QueryFolderContainsFile("credentials")
-	)
-
-	if delete {
-		if folder != nil {
-			if err := os.RemoveAll(folder.Path); err != nil {
-				return nil, errors.Wrap(err, "failed to delete")
-			}
-			fmt.Printf("credentials file in %s has been deleted\n", folder.Path)
-			return nil, nil
-		}
-		fmt.Println("credentials file already inexistant")
-		return nil, nil
-	}
-
-	if noninteractive || reset || folder == nil {
-		readConfig(noninteractive, nostore, &config)
-	} else {
-		data, _ := folder.ReadFile("credentials")
-		json.Unmarshal(data, &config)
-	}
-	return &config, nil
-}
-
 func getBCATransactions(ctx context.Context, bc *bca.BCAApiService, auth []*http.Cookie) ([]bca.Entry, error) {
+	if days > 27 {
+		days = 27
+	}
+	if days < 0 {
+		days = 0
+	}
 	var (
 		end   = time.Now()
-		start = end.AddDate(0, 0, -27)
+		start = end.AddDate(0, 0, -days)
 	)
 	trxs, err := bc.AccountStatementView(ctx, start, end, auth)
 	if err != nil {
@@ -242,95 +228,6 @@ func getBCATransactions(ctx context.Context, bc *bca.BCAApiService, auth []*http
 		return nil, fmt.Errorf("no bca transactions from %s to %s\n", start, end)
 	}
 	return trxs, err
-}
-
-func createYNABTransactions(yc ynab.ClientServicer, trxs []bca.Entry, account *account.Account, budget string) error {
-	ps := make([]transaction.PayloadTransaction, 0)
-	for _, trx := range trxs {
-		ps = append(ps, toPayloadTransaction(trx, account.ID))
-	}
-
-	resp, err := yc.Transaction().CreateTransactions(budget, ps)
-	if err != nil {
-		return err
-	}
-	if len(resp.DuplicateImportIDs) > 0 {
-		fmt.Printf("%d transaction(s) already exists\n", len(resp.DuplicateImportIDs))
-	}
-	fmt.Printf("%d transaction(s) were successfully created\n", len(resp.TransactionIDs))
-	return nil
-}
-
-func getYNABAccount(yc ynab.ClientServicer, budget string, accountName string) (*account.Account, error) {
-	accs, err := yc.Account().GetAccounts(budget, nil)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to get ynab accounts. try -r")
-	}
-
-	for _, acc := range accs.Accounts {
-		if acc.Name == accountName {
-			return acc, nil
-		}
-	}
-	return nil, errors.New("couldnt find account " + accountName)
-}
-
-func createYNABBalancaAdjustment(bc *bca.BCAApiService, ctx context.Context, auth []*http.Cookie, yc ynab.ClientServicer, budget string, a *account.Account) error {
-	bal, err := bc.BalanceInquiry(ctx, auth)
-	if err != nil {
-		return errors.Wrap(err, "failed to get bca balance")
-	}
-	anew, err := yc.Account().GetAccount(budget, a.ID)
-	if err != nil {
-		return errors.Wrap(err, "failed to get ynab account")
-	}
-	delta := bal.Balance.IntPart()*1000 - anew.Balance
-	if delta != 0 {
-		var (
-			payee = "Automated Balance Adjustment"
-		)
-
-		c, err := func() (*category.Category, error) {
-			cs, err := yc.Category().GetCategories(budget, nil)
-			if err != nil {
-				return nil, errors.Wrap(err, "failed to get categories")
-			}
-
-			for _, group := range cs.GroupWithCategories {
-				for _, c := range group.Categories {
-					if c.Name == "Inflows" {
-						return c, nil
-					}
-				}
-			}
-			return nil, errors.New("couldnt find to be budgeted category")
-		}()
-		if err != nil {
-			return err
-		}
-
-		_, err = yc.Transaction().CreateTransaction(budget, transaction.PayloadTransaction{
-			AccountID: a.ID,
-			Date: api.Date{
-				Time: time.Now(),
-			},
-			Amount:     delta,
-			Cleared:    transaction.ClearingStatusReconciled,
-			Approved:   true,
-			PayeeID:    nil,
-			PayeeName:  &payee,
-			CategoryID: &c.ID,
-			Memo:       nil,
-			FlagColor:  nil,
-			ImportID:   nil,
-		})
-		if err != nil {
-			return errors.Wrap(err, "failed to create balance adjustment transaction")
-		}
-
-		fmt.Printf("balance adjustment transaction successfully created\n")
-	}
-	return nil
 }
 
 func transactionsToCsv(trxs []bca.Entry) (string, error) {
@@ -347,47 +244,6 @@ func transactionsToCsv(trxs []bca.Entry) (string, error) {
 
 	trxCsv, err := gocsv.MarshalString(&trxs)
 	return trxCsv, err
-}
-
-func toPayloadTransaction(trx bca.Entry, accountID string) transaction.PayloadTransaction {
-	// description unreliable for hash
-	desc := trx.Description
-	trx.Description = ""
-	// use predicted clearance date for PEND transactions hash
-
-	if trx.Date.IsZero() {
-		trx.Date = clearDate(time.Now())
-	}
-
-	var (
-		t           = trx.Date
-		miliunit    = trx.Amount.Mul(decimal.NewFromInt(1000)).IntPart()
-		payee       = trx.Payee
-		memo        = desc
-		importid, _ = structhash.Hash(trx, 1)
-	)
-	if t.After(time.Now()) {
-		t = time.Now()
-	}
-	if trx.Type == "DB" {
-		miliunit = -miliunit
-	}
-	p := transaction.PayloadTransaction{
-		AccountID: accountID,
-		Date: api.Date{
-			Time: t,
-		},
-		Amount:     miliunit,
-		Cleared:    transaction.ClearingStatusCleared,
-		Approved:   true,
-		PayeeID:    nil,
-		PayeeName:  &payee,
-		CategoryID: nil,
-		Memo:       &memo,
-		FlagColor:  nil,
-		ImportID:   &importid,
-	}
-	return p
 }
 
 // clearDate ref: https://cekmutasi.co.id/news/6/jadwal-jam-cut-off-jam-aktif-mutasi-ibanking
@@ -409,91 +265,4 @@ func clearDate(now time.Time) time.Time {
 		}
 		return rounded.AddDate(0, 0, 1)
 	}
-}
-
-func readConfig(noninteractive, nostore bool, c *config) error {
-	if isZero(c.BCAUser) {
-		if noninteractive {
-			panic(errEmpty)
-		}
-
-		fmt.Print("Enter KlikBCA Username: ")
-		byteUser, _, err := bufio.NewReader(os.Stdin).ReadLine()
-		if err != nil {
-			panic(err)
-		}
-		c.BCAUser = string(byteUser)
-
-		if isZero(c.BCAUser) {
-			panic(errEmpty)
-		}
-	}
-
-	if isZero(c.BCAPassword) {
-		if noninteractive {
-			panic(errEmpty)
-		}
-		fmt.Print("Enter KlikBCA Password: ")
-		bytePassword, err := terminal.ReadPassword(int(syscall.Stdin))
-		if err != nil {
-			panic(err)
-		}
-		c.BCAPassword = string(bytePassword)
-
-		if isZero(c.BCAPassword) {
-			panic(errEmpty)
-		}
-
-		fmt.Println()
-	}
-
-	if isZero(c.YNABToken) {
-		if noninteractive {
-			panic(errEmpty)
-		}
-		fmt.Print("Enter YNAB Personal Access Token: ")
-		byteToken, err := terminal.ReadPassword(int(syscall.Stdin))
-		if err != nil {
-			panic(err)
-		}
-		c.YNABToken = string(byteToken)
-
-		if isZero(c.YNABToken) {
-			panic(errEmpty)
-		}
-
-		fmt.Println()
-	}
-
-	if noninteractive || nostore {
-		return nil
-	}
-
-	// store credentials to user configdir
-	folders := configDirs.QueryFolders(configdir.Global)
-	data, _ := json.Marshal(&c)
-	folders[0].WriteFile("credentials", data)
-	fmt.Printf("saved credentials to %s. use -d to delete or -r to reset anew\n", folders[0].Path)
-
-	return nil
-}
-
-func isZero(i interface{}) bool {
-	return reflect.ValueOf(i).IsZero()
-}
-
-// getPublicIP ref: https://gist.github.com/ankanch/8c8ec5aaf374039504946e7e2b2cdf7f
-func getPublicIP() string {
-	url := "https://api.ipify.org?format=text"
-
-	resp, err := http.Get(url)
-	if err != nil {
-		panic(err)
-	}
-	defer resp.Body.Close()
-	ip, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		panic(err)
-	}
-	return string(ip)
 }
